Import of SGLADKOV/Kafka-0.8001 from CPAN.
gitpan-cpan-distribution: Kafka
gitpan-cpan-version:      0.8001
gitpan-cpan-path:         SGLADKOV/Kafka-0.8001.tar.gz
gitpan-cpan-author:       SGLADKOV
gitpan-cpan-maturity:     released

+++ t/41_fork.t
@@ -42,11 +42,15 @@ use Clone qw(
 );
 use Const::Fast;
 
+use Kafka qw(
+    $RECEIVE_LATEST_OFFSET
+);
 use Kafka::Cluster qw(
     $DEFAULT_TOPIC
 );
 use Kafka::Connection;
 use Kafka::MockIO;
+use Kafka::Consumer;
 use Kafka::Producer;
 
 #-- setting up facilities ------------------------------------------------------
@@ -60,20 +64,97 @@ my $cluster = Kafka::Cluster->new(
 
 #-- declarations ---------------------------------------------------------------
 
-my ( $port, $connection, $topic, $partition, $producer, $response, $is_ready, $pid, $ppid, $success );
+my ( $port, $connection, $topic, $partition, $producer, $response, $consumer, $is_ready, $pid, $ppid, $success, $etalon_messages, $starting_offset );
+
+sub random_strings {
+    my @chars               = ( " ", "A" .. "Z", "a" .. "z", 0 .. 9, qw(! @ $ % ^ & *) );
+    my $msg_len             = 100;
+    my $number_of_messages  = 500;
+
+    note 'generation of messages can take a while';
+    my ( @strings, $size );
+    $strings[ $number_of_messages - 1 ] = undef;
+    foreach my $i ( 0 .. ( $number_of_messages - 1 ) ) {
+        my $len = int( rand( $msg_len ) ) + 1;
+        $strings[ $i ] = join( q{}, @chars[ map { rand( scalar( @chars ) ) } ( 1 .. $len ) ] );
+    }
+    note 'generation of messages complited';
+    return \@strings;
+}
 
 sub sending {
-    return $producer->send(
-        $topic,
-        $partition,
-        'Single message'
-    );
+    my ( $messages ) = @_;
+
+    my $response;
+    eval {
+        foreach my $message ( @$messages ) {
+            undef $response;
+            $response = $producer->send(
+                $topic,
+                $partition,
+                $message
+            );
+        }
+    };
+    fail "sending error: $@" if $@;
+
+    return $response;
+}
+
+sub next_offset {
+    my ( $consumer, $topic, $partition ) = @_;
+
+    my $offsets;
+    eval {
+        $offsets = $consumer->offsets(
+            $topic,
+            $partition,
+            $RECEIVE_LATEST_OFFSET,             # time
+        );
+    };
+    fail "offsets are not received: $@" if $@;
+
+    if( $offsets ) {
+        return $offsets->[0];
+    } else {
+        return;
+    }
 }
 
 sub testing_sending {
-    undef $response;
-    eval { $response = sending() };
-    ++$success unless $@;
+    my $first_offset;
+
+    return
+        unless defined( $first_offset = next_offset( $consumer, $topic, $partition ) );
+    return
+        unless sending( $etalon_messages );
+
+    ++$success;
+
+    return $first_offset;
+}
+
+sub testing_fetching {
+    my ( $first_offset ) = @_;
+
+    my $messages;
+    eval {
+        $messages = $consumer->fetch( $topic, $partition, $first_offset );
+    };
+    fail "messages are not fetched: $@" if $@;
+
+    if ( $messages ) {
+        foreach my $i ( 0 .. $#$etalon_messages ) {
+            my $message = $messages->[ $i ];
+            return unless $message->valid && $message->payload eq $etalon_messages->[ $i++ ];
+        }
+    } else {
+        return;
+    }
+
+    ++$success;
+
+    return $messages;
 }
 
 sub wait_until_ready {
@@ -111,43 +192,70 @@ $connection = Kafka::Connection->new(
 $producer = Kafka::Producer->new(
     Connection  => $connection,
 );
-
-$success = 0;
+$consumer = Kafka::Consumer->new(
+    Connection  => $connection,
+);
 
 # simple sending
-lives_ok { sending() } 'expecting to live';
-ok $success = 1, 'sending OK';
+ok sending( [ 'single message' ] ), 'simple sending ok';
 
-# producer destroyed
 my $clone_connection = clone( $connection );
+# the clients are destroyed
 undef $producer;
 is_deeply( $connection, $clone_connection, 'connection is not destroyed' );
+undef $consumer;
+is_deeply( $connection, $clone_connection, 'connection is not destroyed' );
 
-# recreating producer
+# recreating clients
 $producer = Kafka::Producer->new(
     Connection  => $connection,
 );
-testing_sending();
-ok $success = 2, 'sending OK';
+$consumer = Kafka::Consumer->new(
+    Connection  => $connection,
+);
+
+$success = 0;
+
+$etalon_messages = random_strings();
+$starting_offset = testing_sending();
+ok $success == 1, 'sending OK';
+testing_fetching( $starting_offset );
+ok $success == 2, 'fetching OK';
 
 #-- the producer and connection are destroyed in the child
 
 $is_ready = 0;
 
+$etalon_messages = random_strings();
 if ( $pid = fork ) {                # herein the parent
-    wait_until_ready( 1, $pid );    # expect readiness of the child process
+    $success = 0;
+
     # producer destroyed in the child
-    testing_sending();
-    # $success = 3
+    $starting_offset = testing_sending();
+    # $success == 1
+
+    wait_until_ready( 1, $pid );    # expect readiness of the child process
+
+    testing_fetching( $starting_offset );
+    # $success == 2
     kill 'USR1' => $pid;
 
     wait_until_ready( 2, $pid );    # expect readiness of the child process
+
     # connection destroyed in the child
     $producer = Kafka::Producer->new(
         Connection  => $connection, # potentially destroyed connection
     );
-    testing_sending();
-    # $success = 4
+    $consumer = Kafka::Consumer->new(
+        Connection  => $connection,
+    );
+    # $connection ok
+
+    $etalon_messages = random_strings();
+    $starting_offset = testing_sending();
+    # $success == 3
+    testing_fetching( $starting_offset );
+    # $success == 4
     kill 'USR1' => $pid;
 
     wait;                           # forward to the completion of a child process
@@ -167,13 +275,15 @@ if ( $pid = fork ) {                # herein the parent
     fail 'An unexpected error (fork 1)';
 }
 
-ok $success = 4, 'sending OK';
+ok $success == 4, 'Testing of destruction in the child';
 
 #-- the producer and connection are destroyed in the parent
 
 $is_ready = 0;
 
 if ( $pid = fork ) {                # herein the parent
+    $success = 0;
+
     wait_until_ready( 1, $pid );    # expect readiness of the child process
     undef $producer;
     kill 'USR1' => $pid;
@@ -184,35 +294,51 @@ if ( $pid = fork ) {                # herein the parent
 
     wait;                           # forward to the completion of a child process
 } elsif ( defined $pid ) {          # herein the child process
+    $success = 0;
     $ppid = getppid();
 
-    $success = 0;
+    # producer is not destroyed in the parent
+    $etalon_messages = random_strings();
+    $starting_offset = testing_sending();
+    # $success == 1
+    testing_fetching( $starting_offset );
+    # $success == 2
 
-    testing_sending();
+    $etalon_messages = random_strings();
     kill 'USR1' => $ppid;
-
-    wait_until_ready( 1, $ppid );   # expect readiness of the parent process
     # producer destroyed in the parent
-    testing_sending();
-    # $success = 1
-    kill 'USR1' => $ppid;
+    $starting_offset = testing_sending();
+    # $success == 3
+    testing_fetching( $starting_offset );
+    # $success == 4
+    wait_until_ready( 1, $ppid );   # expect readiness of the parent process
 
+    $etalon_messages = random_strings();
+    kill 'USR1' => $ppid;
     wait_until_ready( 2, $ppid );   # expect readiness of the parent process
+
     # connection destroyed in the parent
     $producer = Kafka::Producer->new(
         Connection  => $connection, # potentially destroyed connection
     );
-    testing_sending();
-    # $success = 2
-    kill 'USR2' => $ppid
-        if $success == 2;
+    $consumer = Kafka::Consumer->new(
+        Connection  => $connection,
+    );
+    # $connection ok
+
+    $starting_offset = testing_sending();
+    # $success == 5
+    testing_fetching( $starting_offset );
+    # $success == 6
+
+    kill 'USR2' => $ppid if $success == 6;  # parent $success increment
 
     exit;
 } else {
     fail 'An unexpected error (fork 2)';
 }
 
-ok $success = 5, 'sending OK';
+ok $success == 1, 'Testing of destruction in the parent';
 
 # POSTCONDITIONS ---------------------------------------------------------------
